Started to use the library code instead of the one in the project to make my life easier.
Extended Glowable with things I think will be useful

Update scene to better showcase the glow effect.

Added the ability to add objects with `Glowable` component to the render list on start if needed.

Added the ability for users to choose if they want the `Glowable` object to be added to the list on start.